+++ ft_putnbrs.c
@@ -6,7 +6,7 @@
 /*   By: bel-idri <bel-idri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 00:24:13 by bel-idri          #+#    #+#             */
-/*   Updated: 2022/11/05 18:26:16 by bel-idri         ###   ########.fr       */
+/*   Updated: 2022/11/07 14:17:58 by bel-idri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	ft_putnbr_base_i_d(int nbr, char *base)
 		nbr *= -1;
 		ft_putnbr_base_i_d(nbr, base);
 	}
-	else if (nbr >= ft_strlen(base))
+	else if (nbr >= (int)ft_strlen(base))
 	{
 		ft_putnbr_base_i_d(nbr / ft_strlen(base), base);
 		ft_putchr(base[nbr % ft_strlen(base)]);
@@ -33,7 +33,7 @@ void	ft_putnbr_base_i_d(int nbr, char *base)
 
 void	ft_putnbr_base_u_x(unsigned int nbr, char *base)
 {
-	if (nbr >= ft_strlen(base))
+	if (nbr >= (unsigned int)ft_strlen(base))
 	{
 		ft_putnbr_base_u_x(nbr / ft_strlen(base), base);
 		ft_putchr(base[nbr % ft_strlen(base)]);
@@ -44,7 +44,7 @@ void	ft_putnbr_base_u_x(unsigned int nbr, char *base)
 
 void	ft_putnbr_base_p(unsigned long long nbr, char *base)
 {
-	if (nbr >= ft_strlen(base))
+	if (nbr >= (unsigned long long)ft_strlen(base))
 	{
 		ft_putnbr_base_p(nbr / ft_strlen(base), base);
 		ft_putchr(base[nbr % ft_strlen(base)]);
